moved fibonacci example into a separate module
